Update tree dir to modular version.
