+++ blendernc/blendernc.py
@@ -4,25 +4,39 @@ from collections import defaultdict
 
 import nodeitems_utils
 
-from . panels import BlenderNC_UI_PT_3dview, BlenderNC_LOAD_OT_On, \
+from blendernc.blendernc.panels import BlenderNC_UI_PT_3dview, BlenderNC_LOAD_OT_On, \
                      BlenderNC_LOAD_OT_Off
 
-from . operators import BlenderNC_OT_ncload, BlenderNC_OT_ncload_Sui, BlenderNC_OT_var, \
+from blendernc.blendernc.operators import BlenderNC_OT_ncload, BlenderNC_OT_ncload_Sui, BlenderNC_OT_var, \
                         BlenderNC_OT_netcdf2img, BlenderNC_OT_preloader,\
                         BlenderNC_OT_apply_material, ImportnetCDFCollection,\
                         Import_OT_mfnetCDF
                         
 
-from . nodes import BlenderNC_NT_netcdf, BlenderNC_NT_preloader,\
-                    BlenderNC_NT_resolution, BlenderNC_NT_output,\
-                    BlenderNC_NT_select_axis, BlenderNC_NT_path, node_categories
+# from . nodes import BlenderNC_NT_netcdf, BlenderNC_NT_preloader,\
+#                     BlenderNC_NT_resolution, BlenderNC_NT_output,\
+#                     BlenderNC_NT_select_axis, BlenderNC_NT_path
+
+from . nodes.node_categories import node_categories
+
+from blendernc.blendernc.nodes.inputs.BlenderNC_NT_path import BlenderNC_NT_path
+from blendernc.blendernc.nodes.inputs.BlenderNC_NT_netcdf import BlenderNC_NT_netcdf
+from blendernc.blendernc.nodes.grid.BlenderNC_NT_resolution import BlenderNC_NT_resolution
+from blendernc.blendernc.nodes.selecting.BlenderNC_NT_select_axis import BlenderNC_NT_select_axis
+from blendernc.blendernc.nodes.selecting.BlenderNC_NT_select_dims import BlenderNC_NT_select_dims
+
+from blendernc.blendernc.nodes.math.BlenderNC_NT_transpose import BlenderNC_NT_transpose
+from blendernc.blendernc.nodes.math.BlenderNC_NT_derivatives import BlenderNC_NT_derivatives
+
+from blendernc.blendernc.nodes.outputs.BlenderNC_NT_output import BlenderNC_NT_output
+from blendernc.blendernc.nodes.outputs.BlenderNC_NT_preloader import BlenderNC_NT_preloader
         
-from . node_tree import create_new_node_tree, BlenderNCNodeTree,\
+from blendernc.blendernc.nodes.node_tree import create_new_node_tree, BlenderNCNodeTree,\
                         node_tree_name
 
-from . sockets import bNCnetcdfSocket,bNCstringSocket
+from blendernc.blendernc.sockets import bNCnetcdfSocket,bNCstringSocket
 
-from .. nodes.cmapsnode import BLENDERNC_CMAPS_NT_node
+from blendernc.blendernc.nodes.cmaps.cmapsnode import BLENDERNC_CMAPS_NT_node
 
 classes = [
     # Panels
@@ -30,11 +44,15 @@ classes = [
     BlenderNC_LOAD_OT_On,
     BlenderNC_LOAD_OT_Off,
     # Nodes
-    BlenderNC_NT_netcdf,
-    BlenderNC_NT_preloader,
-    BlenderNC_NT_resolution,
-    BlenderNC_NT_output,
     BlenderNC_NT_path,
+    BlenderNC_NT_netcdf,
+    BlenderNC_NT_resolution,
+    BlenderNC_NT_select_axis,
+    BlenderNC_NT_select_dims,
+    BlenderNC_NT_transpose,
+    BlenderNC_NT_derivatives,
+    BlenderNC_NT_preloader,
+    BlenderNC_NT_output,
     # Shader Nodes 
     BLENDERNC_CMAPS_NT_node,
     # Operators: files
@@ -47,7 +65,6 @@ classes = [
     BlenderNC_OT_netcdf2img,
     BlenderNC_OT_preloader,
     BlenderNC_OT_apply_material,
-    BlenderNC_NT_select_axis,
     # Sockets
     bNCnetcdfSocket,
     bNCstringSocket,
