+++ blendernc/nodes/cmaps/utils_colorramp.py
